Separate get_diff_stats method to a source file
